refactor(auth): extract bridge result mapping
